+++ src/components/m-header-mobile-navigation/_preview.html
@@ -1,4 +1,4 @@
-<axa-header-mobile-navigation style="min-height: 500px;" offcanvas items='[{"name": "car insurance", "items": [
+<axa-header-mobile-navigation style="min-height: 500px;" items='[{"name": "car insurance", "items": [
        {"name": "professional indemnity insurance"},
        {"name": "public liability insurance"},
        {"name": "shop insurance sub line"},
@@ -72,16 +72,4 @@
        ]}
 
 ]}
-]'>
-  <a href="#" class="m-button m-button--ghost m-button--motion">myAXA Login</a>
-
-  <axa-header-top-content-bar type="commercial">
-    <p class="m-header-top-content-bar__text">commercial <a href="#" class="a-link a-link--white">link</a> stuff</p>
-    <button class="m-button m-button--white m-button--ghost m-header-top-content-bar__button">get it for free</button>
-  </axa-header-top-content-bar>
-
-  <axa-header-meta-mobile-navigation
-    others='[{"name": "individuals", "url": "#", "isActive": true}, {"name": "small busineses", "url": "#"}, {"name": "businesses", "url": "#"}]'
-    languages='[{"code": "de", "url": "#", "isActive": true}, {"code": "en", "url": "#en"}, {"code": "fr", "url": "#fr"}, {"code": "it", "url": "#it"}]'>
-  </axa-header-meta-mobile-navigation>
-</axa-header-mobile-navigation>
+]'></axa-header-mobile-navigation>
