+++ src/components/m-header-mobile/_preview.html
@@ -1,1 +1,99 @@
-<axa-header-mobile></axa-header-mobile>
+<axa-header-mobile offcanvas>
+  <a href="#" class="m-button m-button--ghost m-button--motion">myAXA Login</a>
+
+  <axa-header-mobile-navigation items='[{"name": "car insurance", "items": [
+         {"name": "professional indemnity insurance"},
+         {"name": "public liability insurance"},
+         {"name": "shop insurance sub line"},
+         {"name": "van insurance"},
+         {"name": "residential landlord insurance"}
+         ]},
+  {"name": "home insurance", "items": [
+         {"name": "personal sub line", "items": [
+                {"name": "car insurance"},
+                {"name": "home insurance"},
+                {"name": "travel insurance"},
+                {"name": "breakdown cover"},
+                {"name": "van insurance"}
+         ]},
+         {"name": "business", "items": [
+                {"name": "professional indemnity insurance"},
+                {"name": "public liability insurance"},
+                {"name": "shop insurance"},
+                {"name": "van insurance"},
+                {"name": "residential landlord insurance"}
+         ]}
+  ]},
+  {"name": "health insurance", "items": [
+         {"name": "personal sub line", "items": [
+                {"name": "car insurance"},
+                {"name": "home insurance"},
+                {"name": "travel insurance"},
+                {"name": "breakdown cover"},
+                {"name": "van insurance"}
+         ]},
+         {"name": "business", "items": [
+                {"name": "professional indemnity insurance"},
+                {"name": "public liability insurance"},
+                {"name": "shop insurance"},
+                {"name": "van insurance"},
+                {"name": "residential landlord insurance"}
+         ]}
+
+  ]},
+  {"name": "travel insurance", "items": [
+         {"name": "personal sub line", "items": [
+                {"name": "car insurance"},
+                {"name": "home insurance"},
+                {"name": "travel insurance"},
+                {"name": "breakdown cover"},
+                {"name": "van insurance"}
+         ]},
+         {"name": "business", "items": [
+                {"name": "professional indemnity insurance"},
+                {"name": "public liability insurance"},
+                {"name": "shop insurance"},
+                {"name": "van insurance"},
+                {"name": "residential landlord insurance"}
+         ]}
+
+  ]},
+  {"name": "all products", "items": [
+         {"name": "personal sub line", "items": [
+                {"name": "car insurance"},
+                {"name": "home insurance"},
+                {"name": "travel insurance"},
+                {"name": "breakdown cover"},
+                {"name": "van insurance"}
+         ]},
+         {"name": "business", "items": [
+                {"name": "professional indemnity insurance"},
+                {"name": "public liability insurance"},
+                {"name": "shop insurance"},
+                {"name": "van insurance"},
+                {"name": "residential landlord insurance"}
+         ]}
+
+  ]}
+  ]'></axa-header-mobile-navigation>
+
+  <axa-header-top-content-bar type="commercial">
+    <p class="m-header-top-content-bar__text">commercial <a href="#" class="a-link a-link--white">link</a> stuff</p>
+    <button class="m-button m-button--white m-button--ghost m-header-top-content-bar__button">get it for free</button>
+  </axa-header-top-content-bar>
+
+  <axa-header-mobile-others
+    items='[
+    {"name": "individuals", "url": "#", "isActive": true},
+    {"name": "small busineses", "url": "#"},
+    {"name": "businesses", "url": "#"}
+  ]'
+  ></axa-header-mobile-others>
+
+  <axa-header-mobile-languages items='[
+    {"code": "de", "url": "#", "isActive": true},
+    {"code": "en", "url": "#en"},
+    {"code": "fr", "url": "#fr"},
+    {"code": "it", "url": "#it"}
+  ]'></axa-header-mobile-languages>
+</axa-header-mobile>
