+++ src/components/m-header-mobile/js/header-mobile.js
@@ -6,7 +6,11 @@ class HeaderMobile {
   static DEFAULTS = {
     canvas: '.js-header-mobile-navigation__canvas',
     backdrop: '.js-m-header-mobile-navigation__backdrop',
+    nav: '.js-header-mobile-navigation__nav',
+    category: 'js-header-mobile-navigation__category',
+    back: 'js-header-mobile-navigation__back',
     isMenuOpenClass: 'is-mobile-menu-open',
+    isSubMenuOpenClass: 'is-mobile-sub-menu-open',
     isBackdropFading: 'is-mobile-backdrop-fading',
     isBodyFrozen: 'is-body-frozen',
   }
@@ -18,7 +22,11 @@ class HeaderMobile {
       ...options,
     };
 
+    this.opened = [];
+
     this.handleBackdropClick = this.handleBackdropClick.bind(this);
+    this.handleCategoryClick = this.handleCategoryClick.bind(this);
+    this.handleBackClick = this.handleBackClick.bind(this);
     this.open = this.open.bind(this);
     this.close = this.close.bind(this);
 
@@ -27,6 +35,7 @@ class HeaderMobile {
 
   init() {
     this.canvas = this.rootNode.querySelector(this.options.canvas);
+    this.nav = this.rootNode.querySelector(this.options.nav);
     this.backdrop = this.rootNode.querySelector(this.options.backdrop);
 
     this.on();
@@ -42,11 +51,21 @@ class HeaderMobile {
     this.off();
 
     this.unBackdropClick = on(this.backdrop, 'click', this.handleBackdropClick);
+    this.unCategoryClick = on(this.nav, 'click', this.options.category, this.handleCategoryClick);
+    this.unBackClick = on(this.nav, 'click', this.options.back, this.handleBackClick);
   }
 
   off() {
     if (this.unBackdropClick) {
       this.unBackdropClick();
+    }
+
+    if (this.unCategoryClick) {
+      this.unCategoryClick();
+    }
+
+    if (this.unBackClick) {
+      this.unBackClick();
     }
 
     this.offContextEnabled();
@@ -88,6 +107,15 @@ class HeaderMobile {
 
         // reset initial scroll and menu state
         this.canvas.scrollTop = 0;
+
+        let open = this.opened.pop();
+
+        while (open) {
+          const { parentNode } = open;
+          remove(parentNode, this.options.isSubMenuOpenClass);
+
+          open = this.opened.pop();
+        }
       }
     });
 
@@ -100,13 +128,44 @@ class HeaderMobile {
     publish('header-mobile-navigation/close', null, this._contextNode);
   }
 
+  handleCategoryClick(e, delegateTarget) {
+    e.preventDefault();
+
+    const { parentNode } = delegateTarget;
+
+    if (parentNode.lastChild !== delegateTarget) {
+      const { scrollTop } = this.canvas;
+
+      add(parentNode, this.options.isSubMenuOpenClass);
+
+      this.canvas.scrollTop = 0;
+
+      this.opened.push({
+        parentNode,
+        scrollTop,
+      });
+    }
+  }
+
+  handleBackClick(e) {
+    e.preventDefault();
+
+    const { parentNode, scrollTop } = this.opened.pop();
+
+    remove(parentNode, this.options.isSubMenuOpenClass);
+
+    this.canvas.scrollTop = scrollTop;
+  }
+
   destroy() {
     this.off();
 
     delete this.rootNode;
     delete this.canvas;
+    delete this.nav;
     delete this.backdrop;
     delete this._contextNode;
+    delete this.opened;
   }
 }
