+++ src/components/m-header-mobile/settings/_header-mobile.scss
@@ -8,3 +8,6 @@ $header-mobile__backdrop-z-index: $zindex-header-sub-navigation-backdrop !defaul
 
 $header-mobile__canvas-width-sm: 370px !default;
 $header-mobile__canvas-z-index: $zindex-header-sub-navigation !default;
+
+$header-mobile__item-padding-left: 60px !default;
+$header-mobile__item-padding-right: 40px !default;
